requirements.txt
python-telegram-bot==21.5
feedparser==6.0.11
requests==2.32.3
vaderSentiment==3.3.2
APScheduler==3.10.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
asyncio